Refactor ArchUnit rules
