+++ Testing/src/main/java/com/arivanamin/healthcare/backend/testing/architecture/bases/BaseIntegrationTest.java
@@ -1,0 +1,5 @@
+package com.arivanamin.healthcare.backend.testing.architecture.bases;
+
+public interface BaseIntegrationTest extends BaseUnitTest {
+    
+}
